separated into classes
