--- Project/enemy/firstEnemy.py
@@ -1,1 +1,0 @@
-import pygame
