+++ Project/timeBullet.py
@@ -1,0 +1,30 @@
+import pygame
+from timeConfig import *
+
+class Bullet(pygame.sprite.Sprite):
+    def __init__(self, x, y, direction):
+        pygame.sprite.Sprite.__init__(self)
+        self.speed = 10
+        self.image = pygame.image.load('assets/icons/bullet.png')
+        self.rect = self.image.get_rect()
+        self.rect.center = (x, y)
+        self.direction = direction
+
+    def update(self):
+        #move bullet
+        self.rect.x += (self.direction * self.speed)
+        #check if bullet has gone off screen
+        if self.rect.right < 0 or self.rect.left > SCREEN_WIDTH:
+            self.kill()
+
+        #check collision with characters
+        # if pygame.sprite.spritecollide(player, bullet_group, False):
+        #     if player.alive:
+        #         player.health -= 5
+        #         self.kill()
+        #
+        # 
+        #  if pygame.sprite.spritecollide(enemy, bullet_group, False):
+        #     if enemy.alive:
+        #         enemy.health -= 25
+        #         self.kill()
