+++ Project/timeConfig.py
@@ -1,0 +1,8 @@
+import pygame
+pygame.init()
+
+SCREEN_WIDTH = 900
+SCREEN_HEIGHT = int(SCREEN_WIDTH * 5/9)
+
+bullet_group = pygame.sprite.Group()
+slash_group = pygame.sprite.Group()
